requirements.txt
aiohttp==3.7.4.post0
botbuilder-core==4.14.0
botbuilder-dialogs>=4.7.0
azure-storage==0.35.0
#azure-storage-file-datalake
pandas==1.3.1
pyarrow
azure-identity
azure-keyvault-secrets
pyodbc
